+++ sqlite3_web/extra/js-package/src/channel.ts
@@ -206,9 +206,3 @@ export abstract class ProtocolChannel {
     }
   }
 }
-
-export const channelFinalizationRegistry =
-  FinalizationRegistry &&
-  new FinalizationRegistry<ProtocolChannel>((channel) =>
-    channel._internal_close(),
-  );